pyproject.toml
[project]
name = "contextbase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi>=0.121.1",
    "fastmcp>=2.13.0.2",
    "jsonpointer>=3.0.0",
    "openai>=2.8.0",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]
